requirements.txt
# Core web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Async support
aiofiles==23.2.1

# Google Cloud Storage
google-cloud-storage==2.14.0

# Data validation - Using compatible versions
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development tools (optional)
black==23.12.1
ruff==0.1.11
